requirements.txt
# 核心框架
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-apscheduler==1.13.1

# 爬虫相关
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
lxml==4.9.3

# 数据处理
pandas==2.1.3
numpy==1.24.3

# 任务调度
apscheduler==3.10.4

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志
loguru==0.7.2

# 工具类
python-dateutil==2.8.2
pytz==2023.3 